[refactor] #147 패키지 이동에 따른 변경 진행
